bookmark fixes, renaming
