+++ src/components/bookmark-btn/bookmark-btn.tsx
@@ -1,13 +1,14 @@
 import classNames from 'classnames';
-import { api } from '../../store';
+import { api } from '../../store/store';
 import { useEffect, useState } from 'react';
-import { useAppDispatch, useAppSelector } from '../../hooks';
+import { useAppDispatch, useAppSelector } from '../../hooks/hooks';
 import { incrementFavoritesCount } from '../../store/data/data.slice';
 import { getCityOffers } from '../../store/cities/cities.selectors';
 import { updateCityOffers } from '../../store/cities/cities.slice';
 import { redirectToRouteAction } from '../../store/actions';
 import { APIRoute, AppRoute } from '../../const';
 import { TPlaceCard } from '../place-card/place-card.types';
+import { getIsAuthorized } from '../../store/user/user.selectors';
 
 type BookmarkBtnProps = {
   offerId: TPlaceCard['id'];
@@ -26,18 +27,35 @@ export default function BookmarkBtn({
 }: BookmarkBtnProps): React.ReactNode {
   const [isFav, setIsFav] = useState<boolean>(isFavorite);
   const cityOffers = useAppSelector(getCityOffers);
+  const isAuthorized = useAppSelector(getIsAuthorized);
   const dispatch = useAppDispatch();
 
-  const sizes = {
+  const sizes: { [key in typeof size]: { width: number; height: number } } = {
     small: { width: 18, height: 31 },
     big: { width: 31, height: 33 },
   };
 
   useEffect(() => {
-    setIsFav(isFavorite);
-  }, [isFavorite]);
+    if (isFav !== isFavorite) {
+      setIsFav(isFavorite);
+    }
+  }, [isFavorite, isAuthorized]);
+
+  useEffect(() => {
+    if (!isAuthorized) {
+      setIsFav(false);
+
+      if (onBookmarkDelete) {
+        onBookmarkDelete();
+      }
+    }
+  }, [isAuthorized]);
 
   function handleClick() {
+    if (!isAuthorized) {
+      dispatch(redirectToRouteAction(AppRoute.Login));
+    }
+
     const newStatus = !isFav;
     api
       .post<TPlaceCard>(`${APIRoute.Favorite}/${offerId}/${Number(newStatus)}`)
