+++ src/components/cities-tabs/cities-tabs.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import classNames from 'classnames';
-import { useAppSelector } from '../../hooks';
+import { useAppSelector } from '../../hooks/hooks';
 import { Link, useSearchParams } from 'react-router-dom';
 import { APP_CITIES, SORT_BY_SEARCH_PARAM_NAME } from '../../const';
 import { getCity } from '../../store/cities/cities.selectors';
