+++ src/components/notification/notification.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useAppSelector } from '../../hooks';
+import { useAppSelector } from '../../hooks/hooks';
 import './notification.css';
 import { getError } from '../../store/cities/cities.selectors';
