+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/app/app';
-import { store } from './store';
+import { store } from './store/store';
 import { checkAuthAction } from './store/api-actions';
 
 store.dispatch(checkAuthAction());
