+++ src/pages/offer-page/offer-page.tsx
@@ -8,9 +8,9 @@ import StarsRating from '../../components/stars-rating/stars-rating';
 import { Navigate, useParams } from 'react-router-dom';
 import Map from '../../components/map/map';
 import { ReactNode, useEffect, useState } from 'react';
-import { api } from '../../store';
+import { api } from '../../store/store';
 import Preloader from '../../components/preloader/preloader';
-import { useAppDispatch, useAppSelector } from '../../hooks';
+import { useAppDispatch, useAppSelector } from '../../hooks/hooks';
 import { redirectToRouteAction } from '../../store/actions';
 import { getIsAuthorized } from '../../store/user/user.selectors';
 import BookmarkBtn from '../../components/bookmark-btn/bookmark-btn';
