+++ src/setupTests.ts
@@ -1,6 +1,1 @@
-// import matchers from '@testing-library/jest-dom/matchers';
-// import { expect } from 'vitest';
-
-// expect.extend(matchers);
-
 import '@testing-library/jest-dom';
